pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tts-from-youtube"
version = "0.2.0"
description = "Download YouTube audio, transcribe locally, and re-synthesize clean audio with modern ASR/TTS backends."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "eliujl"}]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "yt-dlp>=2024.10.0",
  "platformdirs>=4.2.0",
  "pydantic>=2.7.0",
  "srt>=3.5.3",
  "regex>=2024.5.15",
]

[project.optional-dependencies]
asr = ["faster-whisper>=1.1.0"]
asr_whisperx = ["whisperx>=3.1.0"]  # optional alignment/diarization (heavy)
tts_piper = ["piper-tts>=1.2.0"]
tts_coqui = ["TTS>=0.22.0; python_version<'3.12'"]
ui = ["gradio>=4.0.0"]
dev = ["ruff>=0.5.0", "pytest>=8.0.0"]

[project.scripts]
y2tts = "tts_from_youtube.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
